Separate 'trainless shares sell for half price' into its own module (same code being used by 1860, 1862, 18GB, this will increase maintainability if there are e.g. any changes needed to bundles_for_corporation in the future
